requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
ollama>=0.1.0
requests>=2.31.0
numpy>=1.24.0
psutil>=5.8.0
GitPython>=3.1.0
flask>=2.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
